Orders.cs: Changed BookId var from int to long

+++ Models/Order.cs
@@ -9,7 +9,7 @@ namespace Project.Models
 	public class Order
 	{
 		public int ID { get; set; }
-		public int BookId { get; set; }
+		public long BookId { get; set; }
 		public int CustomerId { get; set; }
 		public string Address { get; set; }
